fixed a bug with fwf and item use (healing potion on skeleton)
